Move python standalone utilites to _packaging too
