+++ modal/_packaging.py
@@ -1,12 +1,13 @@
 # Copyright Modal Labs 2024
 """Utilities related to packaging dependencies for the container runtime."""
 from pathlib import Path
-from typing import Literal, Tuple
+from typing import Dict, Literal, Tuple
 
 from modal_proto import api_pb2
 from modal_version import __version__
 
 from .config import config
+from .exception import InvalidError
 from .mount import _Mount, module_mount_condition
 
 
@@ -31,3 +32,32 @@ def get_client_package_mounts() -> Tuple[_Mount, _Mount]:
         internal_mount = _Mount.from_name(client_mount_name("internal"), namespace=api_pb2.DEPLOYMENT_NAMESPACE_GLOBAL)
     external_mount = _Mount.from_name(client_mount_name("external"), namespace=api_pb2.DEPLOYMENT_NAMESPACE_GLOBAL)
     return internal_mount, external_mount
+
+
+# Supported releases and versions for python-build-standalone.
+#
+# These can be updated safely, but changes will trigger a rebuild for all images
+# that rely on `add_python()` in their constructor.
+PYTHON_STANDALONE_VERSIONS: Dict[str, Tuple[str, str]] = {
+    "3.8": ("20230826", "3.8.17"),
+    "3.9": ("20230826", "3.9.18"),
+    "3.10": ("20230826", "3.10.13"),
+    "3.11": ("20230826", "3.11.5"),
+    "3.12": ("20240107", "3.12.1"),
+}
+
+
+def python_standalone_mount_name(version: str) -> str:
+    """Get the deployed name of the python-build-standalone mount."""
+    if "-" in version:  # default to glibc
+        version, libc = version.split("-")
+    else:
+        libc = "gnu"
+    if version not in PYTHON_STANDALONE_VERSIONS:
+        raise InvalidError(
+            f"Unsupported standalone python version: {version}, supported values are {list(PYTHON_STANDALONE_VERSIONS.keys())}"
+        )
+    if libc != "gnu":
+        raise InvalidError(f"Unsupported libc identifier: {libc}")
+    release, full_version = PYTHON_STANDALONE_VERSIONS[version]
+    return f"python-build-standalone.{release}.{full_version}-{libc}"
