+++ modal_global_objects/mounts/python_standalone.py
@@ -3,13 +3,10 @@ import shutil
 import tempfile
 import urllib.request
 
+from modal._packaging import python_standalone_mount_name
 from modal.config import config
 from modal.exception import NotFoundError
-from modal.mount import (
-    PYTHON_STANDALONE_VERSIONS,
-    Mount,
-    python_standalone_mount_name,
-)
+from modal.mount import PYTHON_STANDALONE_VERSIONS, Mount
 from modal_proto import api_pb2
